+++ src/vscode/info-message.js
@@ -43,7 +43,7 @@ const reloadMessage =
 
 const installMessage = `Thanks for downloading ${themeTextName} Theme! Use the VSCode Extension's menu to save your preferences and customize your theme! For more information check the theme's page.`;
 
-const updateMessage = `${themeTextName} v2.1.0 is here! Some colors have been changed and a minor update to the minimap. Help support Nebula Oni by reviewing it on the Market Place and help spread the word! Thanks!`;
+const updateMessage = `${themeTextName} v2.1.0 is here! Some colors have been changed, a minor update to the minimap, and added a new Syntax Color: Grey. Help support Nebula Oni by reviewing it on the Market Place and help spread the word! Thanks!`;
 
 const website = `https://github.com/psudo-dev/${themeFileName}`;